move LocationIcon and SlotLabels to components workspace
