+++ components/src/hardware-sim/Deck/index.tsx
@@ -2,4 +2,5 @@ export * from './RobotWorkSpace'
 export * from './RobotCoordsForeignDiv'
 export * from './RobotCoordsForeignObject'
 export * from './RobotCoordsText'
+export * from './SlotLabels'
 export * from './getDeckDefinitions'
