+++ components/src/molecules/index.tsx
@@ -1,1 +1,2 @@
+export * from './LocationIcon'
 export * from './RoundTab'
